v-a-82
# <@&933492547215577155> vote:
## Vote made by <@404014687630524427>
# Summary:
> - Form committees for each department.
- These will be chaired by the Secretaries of State for each department.
- The committee will consist of the holders of the top two offices in the department with additional observer members from the Senate and other departments which they work closely with.
- With a 66% majority the committee can overrule its respective Secretaries of State, they can however still be further overruled by the Senate and of course the Assembly.
# Reasoning:
> - Currently Secretaries of State and ministers are rather distant and often lack strong coordination.
- Currently Secretaries of State have total control over their department, the committee, with direct Senatorial oversight, will serve to counterbalance the power of the Secretaries of State.
- With each department running its own committee it will be easier to assign work to a specific department as the task can be delivered to the committee directly.
# Section of constituition to be changed:
> 3. d)
# To:
```
d. Committee
i. Each department will operate its own Committee, it will be chaired by the respective Secretaries of State.
ii. On each Committee all holders of the senior two most positions within that department, and additional ranks of the department selected to be permitted access will hold seats in the department.
iii. Each Committee will also have observer members, these observer members are external to the department’s normal Committee and will contain all Senators and senior officials from other departments which closely cooperate with each other. These observer members will provide scrutiny, assistance and ideas to support the department in areas it requires assistance.
iv. The Committee will be the main forum for debate on departmental policy and approach, with jobs to be allocated amongst the various members and then sub-allocated to any applicable non-members of the Committee within the department.
v. With a majority of 66% or greater a Committee can overrule and utilise both the individual and collective powers held by the Secretaries of State for that department; they are still however subject to Senatorial overruling.
```
# How to cast your vote:
In order to express support for this change click '👍' on the poll.
In order to dissent to this change click '👎' on the poll.
In order to express your lack of preference to this change click '🤷' on the poll
# List Of Proof Readers:
### <@708099700334657607> with suggestion:
 > Me gusta
### <@339045184300122112>
### <@518403269194678294> with suggestion:
 > E